+++ src/Http/Middleware/Redirect.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Netto\Http\Middleware;
+
+use Closure;
+use Illuminate\Http\Request;
+use Netto\Services\RedirectService;
+use Symfony\Component\HttpFoundation\Response;
+
+class Redirect
+{
+    /**
+     * @param Request $request
+     * @param Closure $next
+     * @return Response
+     */
+    public function handle(Request $request, Closure $next): Response
+    {
+        if ($redirect = RedirectService::request($request)) {
+            return $redirect;
+        }
+
+        return $next($request);
+    }
+}